refact: Move game_statics methods to stat tracker
