+++ runner.rb
@@ -12,6 +12,6 @@ locations = {
 stat_tracker = StatTracker.new
 stat_tracker.from_csv(locations)
 
-game_statistics = GameStatics.new(stat_tracker.games)
-game_statistics.percentage_ties
-game_statistics.count_of_games_by_season
+
+stat_tracker.percentage_ties
+stat_tracker.count_of_games_by_season
