+++ spec/spec_helper.rb
@@ -1,6 +1,5 @@
 require 'simplecov'
 require './lib/game'
-require './lib/game_statics'
 require './lib/team'
 require './lib/game_by_team'
 require './lib/stat_tracker'
